Fixed most memory leaks
There is one more I am aware of but unsure why it is occurring. I have
modified the code to reduce the leak to a fixed amount instead of a
linearly increasing amount. This should fixed the game freezing at uni
computers. Although there is still a chance it might occur the first
time you run the game.